# docs/public/gateway_swagger.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ScriptBee Gateway API",
    "description": "API for ScriptBee Gateway service, managing projects, plugins, and analysis.",
    "version": "v2"
  },
  "servers": [
    {
      "url": "http://localhost:5117/"
    }
  ],
  "paths": {
    "/api/projects": {
      "post": {
        "tags": ["Projects"],
        "summary": "Create a new project",
        "description": "Creates a new project with the specified name and description.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProjectResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Projects"],
        "summary": "Get all projects",
        "description": "Retrieves a list of all existing projects with their basic details.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProjectListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}": {
      "delete": {
        "tags": ["Projects"],
        "summary": "Delete a project",
        "description": "Deletes a project and all its associated data, including scripts and artifacts.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "get": {
        "tags": ["Projects"],
        "summary": "Get project by ID",
        "description": "Retrieves detailed information about a specific project by its unique identifier.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}/scripts": {
      "post": {
        "tags": ["Scripts"],
        "summary": "Create a new script",
        "description": "Creates a new script within the specified project.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScriptCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptData"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Scripts"],
        "summary": "Get all project scripts",
        "description": "Retrieves a list of all scripts available in the specified project.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetScriptDataResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/scripts/{scriptId}": {
      "get": {
        "tags": ["Scripts"],
        "summary": "Get project script by ID",
        "description": "Retrieves metadata about a specific project script.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scriptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptData"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Scripts"],
        "summary": "Update project script metadata",
        "description": "Updates the metadata (name, parameters) of a specific project script.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scriptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScriptCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptData"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}/scripts/{scriptId}/content": {
      "get": {
        "tags": ["Scripts"],
        "summary": "Get project script content",
        "description": "Retrieves the actual code content of a project script.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scriptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Scripts"],
        "summary": "Update project script content",
        "description": "Updates the actual code content of a specific project script.",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scriptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "$ref": "#/components/schemas/Stream"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/plugins/{id}": {
      "get": {
        "tags": ["Plugins"],
        "summary": "Get plugin by ID",
        "description": "Retrieves detailed information about a specific plugin from the marketplace by its unique identifier.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplacePluginWithDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BundleItem": {
        "required": ["id", "name"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "CreateProjectCommand": {
        "required": ["id", "name"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "description": "Command used to create a new project."
      },
      "CreateProjectResponse": {
        "required": ["id", "name", "creationDate"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateScriptCommand": {
        "required": ["path", "language", "parameters"],
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "parameters": {
            "type": ["null", "array"],
            "items": {
              "$ref": "#/components/schemas/ScriptParameter"
            }
          }
        }
      },
      "ExtensionPoint": {
        "required": ["kind"],
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "language": {
            "type": ["null", "string"]
          },
          "extension": {
            "type": ["null", "string"]
          }
        }
      },
      "FileData": {
        "required": ["id", "name"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "GetProjectListResponse": {
        "required": ["data"],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          }
        }
      },
      "GetScriptDataResponse": {
        "required": ["data"],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScriptData"
            }
          }
        }
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": ["null", "string"]
          },
          "title": {
            "type": ["null", "string"]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["null", "integer", "string"],
            "format": "int32"
          },
          "detail": {
            "type": ["null", "string"]
          },
          "instance": {
            "type": ["null", "string"]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "MarketplacePluginWithDetails": {
        "required": ["id", "name", "type", "description", "authors", "versions"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "authors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PluginVersion"
            }
          },
          "bundleItems": {
            "type": ["null", "array"],
            "items": {
              "$ref": "#/components/schemas/BundleItem"
            }
          },
          "sourceCode": {
            "type": ["null", "string"]
          },
          "manifest": {
            "type": ["null", "string"]
          },
          "site": {
            "type": ["null", "string"]
          },
          "license": {
            "type": ["null", "string"]
          },
          "tags": {
            "type": ["null", "array"],
            "items": {
              "type": "string"
            }
          },
          "technologies": {
            "type": ["null", "array"],
            "items": {
              "type": "string"
            }
          },
          "extensionPoints": {
            "type": ["null", "array"],
            "items": {
              "$ref": "#/components/schemas/ExtensionPoint"
            }
          }
        }
      },
      "PluginVersion": {
        "required": ["url", "version", "manifestUrl"],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "manifestUrl": {
            "type": "string"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": ["null", "string"]
          },
          "title": {
            "type": ["null", "string"]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": ["null", "integer", "string"],
            "format": "int32"
          },
          "detail": {
            "type": ["null", "string"]
          },
          "instance": {
            "type": ["null", "string"]
          }
        }
      },
      "Project": {
        "required": ["id", "name", "creationDate"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          }
        },
        "description": "Represents basic project information."
      },
...
      "ProjectDetails": {
        "required": ["id", "name", "creationDate", "savedFiles", "loadedFiles", "linkers"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "savedFiles": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/FileData"
              }
            }
          },
          "loadedFiles": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/FileData"
              }
            }
          },
          "linkers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "description": "Represents detailed project information, including saved and loaded files."
      },
...
      "ScriptData": {
        "required": ["id", "name", "path", "scriptLanguage", "parameters"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "scriptLanguage": {
            "$ref": "#/components/schemas/ScriptLanguage"
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScriptParameter"
            }
          }
        }
      },
      "ScriptLanguage": {
        "required": ["name", "extension"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          }
        }
      },
      "ScriptParameter": {
        "required": ["name", "type", "value"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "value": {}
        }
      },
      "Stream": {
        "type": "string",
        "format": "binary"
      },
...
      "UpdateScriptCommand": {
        "required": ["name", "parameters"],
        "type": "object",
        "properties": {
          "name": {
            "type": ["null", "string"]
          },
          "parameters": {
            "type": ["null", "array"],
            "items": {
              "$ref": "#/components/schemas/ScriptParameter"
            }
          }
        }
      },
...
    }
  },
...
}